third_party/WebKit/LayoutTests/svg/css/svgcursorelement-use-counter.html

<!DOCTYPE html>
<title>SVGCursorElement use counters</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
#target {
  cursor: url(#mycursor), auto;
}
</style>
<script>
// From UseCounter.h
const SVGCursorElement = 1587;
const SVGCursorElementHasClient = 1588;

test(function() {
  assert_false(internals.isUseCounted(document, SVGCursorElement));
  assert_false(internals.isUseCounted(document, SVGCursorElementHasClient));
}, document.title + ', before element parsed');
</script>
<svg><cursor id="mycursor" href="data:image/png;base64,"></cursor></svg>
<script>
test(function() {
  assert_true(internals.isUseCounted(document, SVGCursorElement));
  assert_false(internals.isUseCounted(document, SVGCursorElementHasClient));
}, document.title + ', after element parsed');
</script>
<svg id="target"></svg>
<script>
test(function() {
  document.querySelector('#target').clientLeft;

  assert_true(internals.isUseCounted(document, SVGCursorElement));
  assert_true(internals.isUseCounted(document, SVGCursorElementHasClient));
}, document.title + ', after style computed');
</script>
